backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
apscheduler==3.10.4
openai==1.3.7
anthropic==0.7.8
tweepy==4.14.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1